backend/requirements.txt
absl-py==2.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
asyncpg==0.31.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
deepface==0.0.99
deprecation==2.1.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.0
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.0
fastar==0.11.0
filelock==3.29.0
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fsspec==2026.3.0
gast==0.7.0
gdown==6.0.0
google-pasta==0.2.0
greenlet==3.4.0
grpcio==1.80.0
gunicorn==25.3.0
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.14.0
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.6
lightphe==0.0.24
lz4==4.4.5
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.1
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.7.1
namex==0.1.0
numpy==2.4.4
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==26.1
pandas==3.0.2
pillow==12.2.0
postgrest==2.28.3
propcache==0.4.1
protobuf==7.34.1
psycopg2-binary==2.9.12
pycparser==3.0
pydantic==2.13.2
pydantic-extra-types==2.11.1
pydantic-settings==2.14.0
pydantic_core==2.46.2
Pygments==2.20.0
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.0.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
realtime==2.28.3
requests==2.33.1
retina-face==0.0.17
rich==14.3.4
rich-toolkit==0.19.7
rignore==0.7.6
scikit-learn==1.8.0
scipy==1.17.1
sentry-sdk==2.58.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
starlette==1.0.0
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
sympy==1.14.0
tenacity==9.1.4
tensorflow==2.21.0
termcolor==3.3.0
tf_keras==2.21.0
threadpoolctl==3.6.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.8
wrapt==2.1.2
yarl==1.23.0
zstandard==0.25.0
